+++ Tests/Opc.Ua.Core.Tests/Types/ContentFilter/FilterEvaluatorTests.cs
@@ -401,7 +401,7 @@ namespace Opc.Ua.Core.Tests.Types.ContentFilter
         [Test]
         public void EqualsWithDifferentNumericTypesReturnsFalse()
         {
-            Ua.ContentFilter filter = BuildBinaryFilter(FilterOperator.Equals, Variant.From((int)42), Variant.From((double)42.0));
+            Ua.ContentFilter filter = BuildBinaryFilter(FilterOperator.Equals, Variant.From(42), Variant.From((double)42.0));
             bool result = filter.Evaluate(m_filterContext, m_target);
             Assert.That(result, Is.False);
         }
